feat: Add comprehensive E2E test suite for Codex Dreams feature
This commit introduces a complete end-to-end testing infrastructure for the Codex Dreams
automated insight generation feature, covering all 9 stories in the epic.

## New Test Infrastructure
- `tests/helpers/insights_test_utils.rs` - Test environment setup and utilities
- `tests/helpers/ollama_mock.rs` - Mock Ollama server for predictable testing
- `docs/testing/codex-dreams-e2e-tests.md` - Comprehensive testing documentation
- `scripts/run-e2e-tests.sh` - Automated test runner script

## E2E Test Categories
- **Database Migrations** (`e2e_codex_dreams_migrations.rs`) - Schema evolution and rollback
- **Ollama Integration** (`e2e_ollama_integration.rs`) - LLM client with security/resilience
- **MCP Commands** (`e2e_mcp_insights_commands.rs`) - Model Context Protocol interface
- **Scheduler & Performance** (`e2e_scheduler_performance.rs`) - Background processing
- **Full Pipeline** (`e2e_codex_dreams_full.rs`) - Complete insight generation flow

## Key Features
- Mock Ollama server with configurable failure modes
- Performance benchmarks (>10 memories/sec, <50ms semantic search)
- Security validation (localhost-only connections)
- Comprehensive error handling and resilience testing
- CI/CD integration with feature flag compliance

## Bug Fixes
- Fixed missing fields in Insight struct definitions
- Added missing error variants for codex-dreams feature
- Fixed type consistency issues in auto-tiering engine
- Corrected DatabaseError usage patterns

The test suite ensures production readiness with proper error handling,
performance characteristics, and data integrity guarantees.

+++ src/memory/auto_tiering.rs
@@ -23,7 +23,7 @@ impl AutoTieringEngine {
     }
 
     /// Analyze a memory and determine its appropriate tier
-    pub fn classify_memory(&self, memory: &Memory) -> (MemoryTier, f32) {
+    pub fn classify_memory(&self, memory: &Memory) -> (MemoryTier, f64) {
         let content_lower = memory.content.to_lowercase();
         
         // Suspicious content that indicates data corruption
